% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-multiMixture.R
\docType{methods}
\name{add_model,multi_online_log_mixture-method}
\alias{ANY-method}
\alias{add_model,}
\alias{add_model,multi_online_log_mixture-method}
\title{Method to add an online_log_mixture model}
\usage{
\S4method{add_model}{multi_online_log_mixture}(object, model)
}
\arguments{
\item{object}{A object of class multi_online_log_mixture}

\item{model}{A online_log_mixture model}
}
\value{
An object of type multi_online_log_mixture
}
\description{
Add multiple models to a model comparison object
using the add_model function. Note that for all models
the number of features (p) should be the same.
}
\examples{
M1 <- online_log_mixture(2,1,trace=10)
models <- multi_online_log_mixture(M1)
models <- add_model(models, online_log_mixture(2,2,trace=10))

}

