% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{generate_mixture}
\alias{generate_mixture}
\title{Generate a mixture dataset}
\usage{
generate_mixture(n, p, k, beta = matrix(runif(k * p, -2, 2), nrow = k),
  ak = generate_probability_vector(k))
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of predictors for each individual logistic regression model}

\item{k}{Number of classes}

\item{beta}{A kXp matrix with the true regression coefficients for each mixture component}

\item{ak}{A vector of length k summing to 1 for the mixture probabilties}
}
\value{
A list containing slots \code{data} and \code{params} containing the 
simulated data and the true parameters respectively
}
\description{
This function creates a dataset with n rows, p predictors and k classes,
for simulation of a mixture or logisitc regression models.
}
\examples{
generate_mixture(100, 3, 2)
generate_mixture(100, 2, 1)
generate_mixture(10^4, 4,6)
}
\keyword{generate}
\keyword{logistic}
\keyword{mixture}

