% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{generate_probability_vector}
\alias{generate_probability_vector}
\title{Genererate a probablity vector}
\usage{
generate_probability_vector(k)
}
\arguments{
\item{k}{Number of components}
}
\value{
Vector of lenght \code{k} with random entries that sum to 1
}
\description{
Utility function to generate a probalbity vector of specific lenght
}

