% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Compute inverse logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numeric vector (values between -Inf and Inf)}
}
\value{
Numberic vector with lenght \code{lenght(x)} with entries between 0 and 1
}
\description{
Utility function to compute an inverse logit
}

