% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-onlineLogMixture.R
\name{ll_compute}
\alias{ll_compute}
\title{(Internal) function to compute the log likelihood of an observation}
\usage{
ll_compute(y, X, beta, ak = 1, wk = 1)
}
\arguments{
\item{y}{A scalar with value 0 or 1; the dependent variable}

\item{X}{The feature vector of the current observation}

\item{beta}{A $k$ by $p$ matrix with the true regression coefficients for each mixture component}

\item{ak}{A vector of length k summing to 1 for the mixture probabilties}

\item{wk}{The mixture weights (posterior probablity of the data point for each component)}
}
\value{
A list containing the log likelihood of the datapoint and
the max log likelihood of the datapoint
}
\description{
Function takes the current model parameters, and an observation to
compute the log likelihood of the observation
}

