% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{ofmlr-package}
\alias{ofmlr}
\alias{ofmlr-package}
\title{Fit online finite mixtures of logistic regression models.}
\description{
mixtures of logistic regression models in parallel. The main workhorse \code{online_log_mixture()} 
and its utility functions (such as \code{add_observation()} can be used to fit a model. 
The \code{multi_online_log_mixture} object allows one to fit multiple models in paralel
and compare their outcomes.
}

