% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-onlineLogMixture.R
\name{online_log_mixture}
\alias{online_log_mixture}
\title{Create an S4 object to fit the online logistic mixture model}
\usage{
online_log_mixture(p, k, beta = matrix(runif(k * p, -1, 1), nrow = k),
  ak = generate_probability_vector(k), ll.window = 500, trace = FALSE)
}
\arguments{
\item{p}{Number of predictors for each logistic regression model}

\item{k}{Number of mixture components}

\item{beta}{A $k$ by $p$ matrix with the true regression coefficients for each mixture component}

\item{ak}{A vector of length $k$ summing to 1 for the mixture probabilties}

\item{ll.window}{The length of the (lagged) mean log-likelihood window}

\item{trace}{Whether or not the parameter estimates and the lagged mean log likelihood should be traced
can be set to an integer, $x > 0$, and it will store a snapshot every x-th datapoint.
WARINING: when using \code{trace=1} the object grows large very quickly}
}
\value{
An object of class "online_log_mixture"
The object contains the following slots:
\describe{
   \item{\code{params}}{A list containing all the parameters of the model. Contains the objects \code{beta}, \code{Sak}, \code{n}}
   \item{\code{descriptives}}{A list containing the (lagged) mean log-likelihood, the max log-likelihood, and the AIC and BIC approximations}
   \item{\code{trace}}{A list containing the trace of each of the parameters}
 }
}
\description{
This function allows you to initialize an online logistic regression
model. After initialization by providing the model description (in terms of 
the number of predictors $p$ of each logistic regression model and the number
of mixture components $k$) one can use the \code{\link{add.observation}}
method to add observations and update the parameters and call the \code{\link{summary()}} and \code{\link{plot()}}
methods.
}
\examples{
online_log_mixture(3,2)
online_log_mixture(2,4, ak=c(.1,.2,.3,.4), trace=100)
}
\seealso{
For more detailed examples see \code{\link{add_observation}}
}

