% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-multiMixture.R
\docType{methods}
\name{plot,multi_online_log_mixture,missing-method}
\alias{plot,multi_online_log_mixture,missing-method}
\title{Plot method for the mult_online_log_mixture class}
\usage{
\S4method{plot}{multi_online_log_mixture,missing}(x, y, params = FALSE,
  omit = 100, param.y = c(-5, 5), ...)
}
\arguments{
\item{x}{An object of type multi_online_log_mixture}

\item{y}{NULL}

\item{params}{Boolean, if TRUE the trace of the parameter values will also be printed}

\item{omit}{Number of observations to omit from the log likelihood and l2 Norm traces}

\item{.y}{a vector with the min and max values of the plot of the beta parameters}
}
\description{
Will create a plot of each of the models stored in 
the model comparison class that you can browse one by
one.
}
\examples{
M1 <- online_log_mixture(2,1)
models <- multi_online_log_mixture(M1)
models <- add_model(models, online_log_mixture(2,2))
models <- add_observation(models, 1, c(2,-3))
plot(models, params=TRUE)

}

