% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-onlineLogMixture.R
\docType{methods}
\name{plot,online_log_mixture,missing-method}
\alias{plot,online_log_mixture,missing-method}
\title{Plot method for the online_log_mixture class}
\usage{
\S4method{plot}{online_log_mixture,missing}(x, y, params = FALSE,
  omit = 100, param.y = c(-5, 5), ...)
}
\arguments{
\item{x}{The online_log_mixture object}

\item{y}{NULL}

\item{params}{Boolean, if TRUE the trace of the parameter values will also be printed}

\item{omit}{Number of observations to omit from the log likelihood and l2 Norm traces}

\item{.y}{a vector with the min and max values of the plot of the beta parameters}
}
\description{
Plot an object of type online_log_mixture.
This will only produce a plot when \code{trace!=FALSE}
The plots will be of the log-likelihood of the model over
the number of observations and the average change in L2
norm of the model parameters. Also, when \code{params=TRUE}
plots of the parameter estimates over time will also be produced.
}
\examples{
M2 <- online_log_mixture(3,3, trace=1)
for(i in 1:10000){ 	
	X <- runif(3,-2,2)
	y <- rbinom(1, 1, inv_logit(c(0,-2,2)\%*\%X))
	M2 <- add_observation(M2, y, X, 0)
}
plot(M2, params=TRUE)

}

