% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-onlineLogMixture.R
\docType{methods}
\name{summary,online_log_mixture-method}
\alias{summary,online_log_mixture-method}
\title{Summary method for the online_log_mixture class}
\usage{
\S4method{summary}{online_log_mixture}(object, ...)
}
\arguments{
\item{object}{The fitted model}
}
\value{
Nothing
}
\description{
Takes an object of type online_log_mixture and prints
the model parameters and a number of descriptives
}

